transcription_service/requirements.txt
google-generativeai
Pillow
pillow_heif
python-dotenv
